+++ src/Discord/Internal/Rest/Channel.hs
@@ -106,7 +106,7 @@ data MessageDetailedOpts = MessageDetailedOpts
   , messageDetailedFile                     :: Maybe (T.Text, B.ByteString)
   , messageDetailedAllowedMentions          :: Maybe AllowedMentions
   , messageDetailedReference                :: Maybe MessageReference
-  } deriving (Show, Eq, Ord)
+  } deriving (Show, Read, Eq, Ord)
 
 instance Default MessageDetailedOpts where
   def = MessageDetailedOpts { messageDetailedContent         = ""
@@ -125,7 +125,7 @@ data AllowedMentions = AllowedMentions
   , mentionUserIds     :: [UserId]
   , mentionRoleIds     :: [RoleId]
   , mentionRepliedUser :: Bool
-  } deriving (Show, Eq, Ord)
+  } deriving (Show, Read, Eq, Ord)
 
 instance Default AllowedMentions where
   def = AllowedMentions { mentionEveryone    = False
@@ -152,7 +152,7 @@ instance ToJSON AllowedMentions where
 data ReactionTiming = BeforeReaction MessageId
                     | AfterReaction MessageId
                     | LatestReaction
-  deriving (Show, Eq, Ord)
+  deriving (Show, Read, Eq, Ord)
 
 reactionTimingToQuery :: ReactionTiming -> R.Option 'R.Https
 reactionTimingToQuery t = case t of
@@ -165,7 +165,7 @@ data MessageTiming = AroundMessage MessageId
                    | BeforeMessage MessageId
                    | AfterMessage MessageId
                    | LatestMessages
-  deriving (Show, Eq, Ord)
+  deriving (Show, Read, Eq, Ord)
 
 messageTimingToQuery :: MessageTiming -> R.Option 'R.Https
 messageTimingToQuery t = case t of
@@ -179,7 +179,7 @@ data ChannelInviteOpts = ChannelInviteOpts
   , channelInviteOptsMaxUsages              :: Maybe Integer
   , channelInviteOptsIsTemporary            :: Maybe Bool
   , channelInviteOptsDontReuseSimilarInvite :: Maybe Bool
-  } deriving (Show, Eq, Ord)
+  } deriving (Show, Read, Eq, Ord)
 
 instance ToJSON ChannelInviteOpts where
   toJSON ChannelInviteOpts{..} = object [(name, val) | (name, Just val) <-
@@ -197,7 +197,7 @@ data ModifyChannelOpts = ModifyChannelOpts
   , modifyChannelUserRateLimit        :: Maybe Integer
   , modifyChannelPermissionOverwrites :: Maybe [Overwrite]
   , modifyChannelParentId             :: Maybe ChannelId
-  } deriving (Show, Eq, Ord)
+  } deriving (Show, Read, Eq, Ord)
 
 instance ToJSON ModifyChannelOpts where
   toJSON ModifyChannelOpts{..} = object [(name, val) | (name, Just val) <-
@@ -214,11 +214,11 @@ data ChannelPermissionsOpts = ChannelPermissionsOpts
   { channelPermissionsOptsAllow :: Integer
   , channelPermissionsOptsDeny :: Integer
   , channelPermissionsOptsType :: ChannelPermissionsOptsType
-  } deriving (Show, Eq, Ord)
+  } deriving (Show, Read, Eq, Ord)
 
 data ChannelPermissionsOptsType = ChannelPermissionsOptsUser
                                 | ChannelPermissionsOptsRole
-  deriving (Show, Eq, Ord)
+  deriving (Show, Read, Eq, Ord)
 
 instance ToJSON ChannelPermissionsOptsType where
   toJSON t = case t of ChannelPermissionsOptsUser -> String "member"
@@ -234,7 +234,7 @@ data GroupDMAddRecipientOpts = GroupDMAddRecipientOpts
   { groupDMAddRecipientUserToAdd :: UserId
   , groupDMAddRecipientUserToAddNickName :: T.Text
   , groupDMAddRecipientGDMJoinAccessToken :: T.Text
-  } deriving (Show, Eq, Ord)
+  } deriving (Show, Read, Eq, Ord)
 
 channelMajorRoute :: ChannelRequest a -> String
 channelMajorRoute c = case c of