+++ src/Discord/Internal/Rest/Guild.hs
@@ -144,7 +144,7 @@ data ModifyGuildIntegrationOpts = ModifyGuildIntegrationOpts
   { modifyGuildIntegrationOptsExpireBehavior :: Integer
   , modifyGuildIntegrationOptsExpireGraceSeconds :: Integer
   , modifyGuildIntegrationOptsEmoticonsEnabled :: Bool
-  } deriving (Show, Eq, Ord)
+  } deriving (Show, Read, Eq, Ord)
 
 instance ToJSON ModifyGuildIntegrationOpts where
   toJSON ModifyGuildIntegrationOpts{..} =  object [(name, val) | (name, Just val) <-
@@ -154,7 +154,7 @@ instance ToJSON ModifyGuildIntegrationOpts where
 
 data CreateGuildIntegrationOpts = CreateGuildIntegrationOpts
   { createGuildIntegrationOptsType :: T.Text
-  } deriving (Show, Eq, Ord)
+  } deriving (Show, Read, Eq, Ord)
 
 instance ToJSON CreateGuildIntegrationOpts where
   toJSON CreateGuildIntegrationOpts{..} =  object [(name, val) | (name, Just val) <-
@@ -163,7 +163,7 @@ instance ToJSON CreateGuildIntegrationOpts where
 data CreateGuildBanOpts = CreateGuildBanOpts
   { createGuildBanOptsDeleteLastNMessages :: Maybe Int
   , createGuildBanOptsReason              :: Maybe T.Text
-  } deriving (Show, Eq, Ord)
+  } deriving (Show, Read, Eq, Ord)
 
 instance ToJSON CreateGuildBanOpts where
   toJSON CreateGuildBanOpts{..} =  object [(name, val) | (name, Just val) <-
@@ -177,7 +177,7 @@ data ModifyGuildRoleOpts = ModifyGuildRoleOpts
   , modifyGuildRoleOptsColor           :: Maybe ColorInteger
   , modifyGuildRoleOptsSeparateSidebar :: Maybe Bool
   , modifyGuildRoleOptsMentionable     :: Maybe Bool
-  } deriving (Show, Eq, Ord)
+  } deriving (Show, Read, Eq, Ord)
 
 instance ToJSON ModifyGuildRoleOpts where
   toJSON ModifyGuildRoleOpts{..} =  object [(name, val) | (name, Just val) <-
@@ -193,7 +193,7 @@ data AddGuildMemberOpts = AddGuildMemberOpts
   , addGuildMemberOptsRoles       :: Maybe [RoleId]
   , addGuildMemberOptsIsMuted     :: Maybe Bool
   , addGuildMemberOptsIsDeafened  :: Maybe Bool
-  } deriving (Show, Eq, Ord)
+  } deriving (Show, Read, Eq, Ord)
 
 instance ToJSON AddGuildMemberOpts where
   toJSON AddGuildMemberOpts{..} =  object [(name, val) | (name, Just val) <-
@@ -209,7 +209,7 @@ data ModifyGuildMemberOpts = ModifyGuildMemberOpts
   , modifyGuildMemberOptsIsMuted       :: Maybe Bool
   , modifyGuildMemberOptsIsDeafened    :: Maybe Bool
   , modifyGuildMemberOptsMoveToChannel :: Maybe ChannelId
-  } deriving (Show, Eq, Ord)
+  } deriving (Show, Read, Eq, Ord)
 
 instance ToJSON ModifyGuildMemberOpts where
   toJSON ModifyGuildMemberOpts{..} =  object [(name, val) | (name, Just val) <-
@@ -229,7 +229,7 @@ data CreateGuildChannelOpts
   , createGuildChannelOptsMaxUsers :: Maybe Integer
   , createGuildChannelOptsCategoryId :: Maybe ChannelId }
   | CreateGuildChannelOptsCategory
-  deriving (Show, Eq, Ord)
+  deriving (Show, Read, Eq, Ord)
 
 createChannelOptsToJSON :: T.Text -> [Overwrite] -> CreateGuildChannelOpts -> Value
 createChannelOptsToJSON name perms opts = object [(key, val) | (key, Just val) <- optsJSON]
@@ -266,7 +266,7 @@ data ModifyGuildOpts = ModifyGuildOpts
    -- VerificationLevel
    -- DefaultMessageNotification
    -- ExplicitContentFilter
-  } deriving (Show, Eq, Ord)
+  } deriving (Show, Read, Eq, Ord)
 
 instance ToJSON ModifyGuildOpts where
   toJSON ModifyGuildOpts{..} =  object [(name, val) | (name, Just val) <-
@@ -278,7 +278,7 @@ instance ToJSON ModifyGuildOpts where
 data GuildMembersTiming = GuildMembersTiming
                           { guildMembersTimingLimit :: Maybe Int
                           , guildMembersTimingAfter :: Maybe UserId
-                          } deriving (Show, Eq, Ord)
+                          } deriving (Show, Read, Eq, Ord)
 
 guildMembersTimingToQuery :: GuildMembersTiming -> R.Option 'R.Https
 guildMembersTimingToQuery (GuildMembersTiming mLimit mAfter) =