+++ src/Discord/Internal/Rest/User.hs
@@ -54,7 +54,7 @@ data UserRequest a where
 
 -- | Formatted avatar data https://discord.com/developers/docs/resources/user#avatar-data
 data CurrentUserAvatar = CurrentUserAvatar T.Text
-  deriving (Show, Eq, Ord)
+  deriving (Show, Read, Eq, Ord)
 
 parseCurrentUserAvatar :: B.ByteString -> Either T.Text CurrentUserAvatar
 parseCurrentUserAvatar bs =