+++ changelog.md
@@ -6,6 +6,8 @@ Discord API changes, so use the most recent version at all times
 
 ## master
 
+[drewolson](https://github.com/aquarial/discord-haskell/pull/80) allows parsing an optional guild region
+
 ## 1.9.1
 
 Add [color attribute for CreateEmbed](https://github.com/aquarial/discord-haskell/issues/78)